+++ src/software/util/make_enum/make_enum.hpp
@@ -1,6 +1,7 @@
 #pragma once
 
 #include <stdexcept>
+#include <string>
 
 #include "software/util/make_enum/make_enum_helpers.hpp"
 
@@ -64,7 +65,7 @@
         /* This index lookup relies on the assumption that the enum does not manually */   \
         /* specify any values. If it did, the underlying integer of the given value */     \
         /* may be out of range of the vector of strings */                                 \
-        os << reflective_enum::valueNames<name>().at(static_cast<int>(value));             \
+        os << reflective_enum::valueNames<name>().at(static_cast<std::size_t>(value));             \
         return os;                                                                         \
     }
